pyproject.toml
[project]
name = "smarttasker"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "aiosqlite>=0.20.0",
  "bs4>=0.0.2",
  "flet==0.25.2",
  "google-genai>=0.6.0",
  "pytest>=8.3.4",
  "pytest-asyncio>=0.25.2",
  "python-dotenv>=1.0.1",
  "sqlalchemy[asyncio]>=2.0.37",
]


[tool.flet]
# org name in reverse domain name notation, e.g. "com.mycompany".
# Combined with project.name to build bundle ID for iOS and Android apps
org = "com.binarycrafts"

# project display name that is used as an app title on Android and iOS home screens,
# shown in window titles and about app dialogs on desktop.
product = "smarttasker"

# company name to display in about app dialogs
company = "Flet"

# copyright text to display in about app dialogs
copyright = "Copyright (C) 2024 by Flet"

[tool.flet.app]
path = "src"

[tool.uv]
dev-dependencies = [
    "flet[all]==0.25.2",
]

[tool.flet.splash]
android = false

[tool.flet.android]
split_per_abi = true
